requirements.dev.txt
ipdb
ipython
pudb
pytest-watch
flake8
pyproject-flake8
black
isort
mkdocs
wheel
twine
mypy
sqlalchemy==1.4.35
sqlmodel